+++ lib/core.dart
@@ -7,11 +7,15 @@
 library core;
 
 import "dart:io";
+import "dart:mirrors";
+
+import 'annotation.dart';
 
 part 'src/core/application.dart';
-part 'src/core/config.dart';
-part 'src/core/runtime_delegate.dart';
+part 'src/core/rest_runtime.dart';
 part 'src/core/rest_exception.dart';
 part 'src/core/exception_mapper.dart';
 part 'src/core/application_context.dart';
 part 'src/core/service_context.dart';
+part 'src/core/processor.dart';
+part 'src/core/parser.dart';