+++ lib/src/annotation/annotation.dart
@@ -1,0 +1,11 @@
+//Copyright (C) 2013 Sergey Akopkokhyants. All Rights Reserved.
+//Author: akserg
+
+part of annotation;
+
+/**
+ * Abstract annotation class is the base for other annotations.
+ */
+class Annotation {
+  const Annotation();
+}