+++ lib/src/core/config.dart
@@ -1,4 +1,0 @@
-//Copyright (C) 2013 Sergey Akopkokhyants. All Rights Reserved.
-//Author: akserg
-
-part of core;