+++ changes.md
@@ -5,7 +5,8 @@ Restful Changes
 
 * Created annotation and core libraries.
 * Created example not finished yet.
-* Added short version of annotation GET, HEAD, DELETE, POST and PUT
+* Added short version of annotation GET, HEAD, DELETE, POST and PUT.
+* Added ApplicationContext and ServiceContext. Class RuntimeDelegate renamed to RestRuntime.
 
 ** September 9, 2013